+++ src/app/(route)/goal/editor/[id]/repeat/components/DDuDuRepeatEditor/hooks/index.ts
@@ -1,0 +1,1 @@
+export { default as useRepeatEditor } from "./useRepeatEditor/useRepeatEditor";
